+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.1.1 (November 21, 2017)
+- added possibility to add props to form element via [formProps](https://github.com/cat-react/form/blob/master/docs/api.md#formprops) prop
+- the validation rule `isRequired` now checks for `undefined`, `null` or an empty string. everything else is valid
+
 ### 0.1.0 (October 3, 2017)
 - first official release!
 - added api documentation